PROJECT RESTRUCTURE
